Refactor sorting util functions
